tests: refactored getColorForShowIds test. moved util tests to util dir. downgraded ts-jest. did minor cleanup
