+++ front/jest.config.js
@@ -1,11 +1,6 @@
 module.exports = {
   roots: ['<rootDir>/src'],
-
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest',
-  },
-
+  transform: { '^.+\\.tsx?$': 'ts-jest' },
   testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$',
-
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
 };
