+++ front/src/utils/__tests__/maybePluralize.test.ts
@@ -1,13 +1,12 @@
-import { maybePluralize } from '../../utils/formatting';
+import { maybePluralize } from '../formatting';
 
 describe('maybePluralize', () => {
-
   it('pluralizes a string', () => {
     const pluralized = maybePluralize(2, 'test');
     expect(pluralized).toBe('tests');
   });
-  
-  it('does not pluralizes a string', () => {
+
+  it('does not pluralize a string', () => {
     const pluralized = maybePluralize(1, 'test');
     expect(pluralized).toBe('test');
   });
